20260316/requirements.txt
sentence-transformers
faiss-cpu
rank-bm25
stable-baselines3
gymnasium
networkx
opentelemetry-api
opentelemetry-sdk
pyyaml